# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "ToDo API",
    "description": "API for a ToDo App to remember tasks ",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api-key",
      "in": "header"
    }
  },
  "paths": {
    "/tasks": {
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get all tasks",
        "description": "Get all tasks from the json file and displays them",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Json displayed",
            "schema": {
              "$ref": "#/definitions/tasks"
            }
          },
          "404": {
            "description": "json not found   no tasks available"
          }
        }
      },
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Post a new tasks",
        "description": "Post a new tasks as a json object",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "type": "string"
          },
          {
            "name": "description",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Task posted",
            "schema": {
              "$ref": "#/definitions/tasks"
            }
          },
          "400": {
            "description": "some values are null"
          }
        }
      }
    }
  },
  "definitions": {
    "server side error": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string",
          "example": "ERROR"
        },
        "msg": {
          "type": "string",
          "example": "some error message"
        },
        "error": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "Error message caught"
            },
            "name": {
              "type": "string",
              "example": "Error name"
            },
            "stack": {
              "type": "string",
              "example": "Error stack"
            }
          },
          "required": [
            "message",
            "name"
          ]
        }
      },
      "required": [
        "status",
        "msg"
      ]
    },
    "tasks": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 0
        },
        "createdTateTime": {
          "type": "string",
          "example": "now"
        },
        "finishedDateTime": {},
        "title": {
          "type": "string",
          "example": "Create some random data"
        },
        "description": {
          "type": "string",
          "example": "Create some random data for the documentation"
        },
        "creatorId": {
          "type": "number",
          "example": 0
        }
      },
      "required": [
        "id",
        "createdTateTime",
        "finishedDateTime",
        "title",
        "description",
        "creatorId"
      ]
    },
    "lends": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 100
        },
        "customer_id": {
          "type": "number",
          "example": 123
        },
        "isbn": {
          "type": "number",
          "example": 402
        },
        "borrowed_at": {
          "type": "string",
          "example": "2023-06-14"
        },
        "returned_at": {}
      },
      "required": [
        "id",
        "customer_id",
        "isbn",
        "borrowed_at",
        "returned_at"
      ]
    }
  }
}